split condition_variable.cpp on two files
Move GCC related code from condition_variable.cpp to
condition_variable.cc. The .cc file is placed together with
mutex and future GCC files. In other words, keep GCC
files in its own directory.
